+++ Sources/Datastore/CoreData/EntityRecord.swift
