+++ Sources/Datastore/CoreData/Properties/DoubleProperty.swift
