+++ Sources/Datastore/CoreData/Properties/NamedProperty.swift
