+++ Sources/Datastore/CoreData/Properties/StringProperty.swift
