--- Sources/Datastore/Model/CombinedEntity.swift
@@ -1,11 +1,0 @@
-// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-//  Created by Sam Deane on 17/12/2019.
-//  All code (c) 2019 - present day, Elegant Chaos Limited.
-// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-
-import Foundation
-
-struct CombinedEntity {
-    let reference: EntityReference
-    let properties: PropertyDictionary
-}
