+++ Sources/Datastore/Resolvers/NullResolver.swift
